+++ lib/solidus_weighted_shipping/domain.rb
@@ -3,6 +3,7 @@
 require "solidus_weighted_shipping/errors"
 require "solidus_weighted_shipping/decimal"
 require "solidus_weighted_shipping/rate_table"
+require "solidus_weighted_shipping/legacy_preferences"
 require "solidus_weighted_shipping/package_input"
 require "solidus_weighted_shipping/constraints"
 require "solidus_weighted_shipping/quote"